artefacts/ndci_sentinel2/requirements.txt
# Runtime
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9       # PostgreSQL driver (remova se usar só SQLite)
alembic>=1.13.0              # migrações (opcional — as SQLs em migrations/ são suficientes)

# GEE
earthengine-api>=0.1.380
google-auth>=2.28.0
google-auth-httplib2>=0.2.0

# ML
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0

# Imagens (para export GIF/ZIP — opcional)
Pillow>=10.3.0
